refactor(exec): enhance memory management and modularize code
Added new memory management functions with logging capabilities
(enabled when the development macro is set), and integrated checks
to ensure proper allocation. These changes improve memory safety
and traceability.

Addressed the issue of code concentration by partially refactoring
and modularizing related functions, improving maintainability.

No significant changes were made to execution behavior (pipeline,
background processes, redirection, operators), and known issues
remain unresolved but are slated for upcoming fixes.

Refs: #6
