+++ error.h
@@ -1,0 +1,6 @@
+#ifndef __RICKSHELL_ERROR_H__
+#define __RICKSHELL_ERROR_H__
+#include <stdio.h>
+
+void print_error(const char* message);
+#endif /* __RICKSHELL_ERROR_H__ */
