+++ pipeline.h
@@ -1,0 +1,6 @@
+#ifndef __RICKSHELL_PIPELINE_H__
+#define __RICKSHELL_PIPELINE_H__
+#include "expr.h"
+
+int execute_pipeline(Command* cmd);
+#endif /* __RICKSHELL_PIPELINE_H__ */
