+++ rick.h
@@ -1,0 +1,5 @@
+#ifndef __RICKSHELL_RICK_H__
+#define __RICKSHELL_RICK_H__
+typedef struct Redirect Redirect;
+typedef enum   RedirectType RedirectType;
+#endif
